+++ MelonLoader/Extensions/Fixes/QuitFix.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.Runtime.CompilerServices;
 
 namespace MelonLoader.Fixes
 {