+++ CHANGELOG.md
@@ -131,6 +131,7 @@
 106. Converted Assembly Generator to be ran by Base Assembly.
 107. Fixed Issue with Parsing Unity Version on Weirdly Marked Versions.
 108. Fixed Issue with Mono Base Directory being Wrong causing Crashes.
+109. Added Warning for Transpiler Use on Il2Cpp Unhollowed Methods.
 
 ---